separate to MVC in Controller file
